backend/requirements.txt
flask>=3.0
flask-cors>=4.0
python-dotenv>=1.0
requests>=2.31
langchain>=0.2
langchain-community>=0.2
langchain-core>=0.2
chromadb>=0.5
pypdf>=4.0
werkzeug>=3.0